refactor NewAgentRequest func
